+++ CHANGELOG.md
@@ -3,6 +3,12 @@ QUEASARS change log
 
 Current version: 0.2.0
 
+## Unreleased
+
+### Added
+
+- Added tournament selection as an alternative selection method for EVQE ([Issue #25])
+
 ## 0.2.0
 
 - Implement a general algorithm structure for Evolving Ansatz VQE algorithms
@@ -12,3 +18,5 @@ Current version: 0.2.0
 ## 0.1.0
 
 - Initial codeless pypi commit
+
+[Issue #25]: https://github.com/DLR-RB/QUEASARS/issues/25